backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
websockets==12.0
pydantic==2.5.3
python-multipart==0.0.6

# Performance optimizations
orjson==3.9.15

# Optional: ML Libraries (uncomment as needed)
# tensorflow==2.15.0
# torch==2.1.2
# adversarial-robustness-toolbox==1.16.0
# numpy==1.26.3
# Pillow==10.2.0
